fabtests/common add option to configure message sizes
The change adds an option to specify a range of messages
to fabtests by using the format 'r:start,increment,end'
to the "-S" option. Prior to this change, we were able to
provide "all" or a specific size to the message size option.

Updates have been made to pytest/efa tests to use this option
for message sizes in the range of 8k-9k with an increment of 4.

Tests were conducted using runfabtests.py to ensure the message
ranges are being processed by pytest (cuda testing is not done).

Signed-off-by: mmubarak <misbah.mubarak@gmail.com>